:recycle: owasm: Refactor runtime code into owasm package
