+++ go-owasm/Cargo.toml
@@ -8,13 +8,7 @@ edition = "2018"
 crate-type = ["cdylib"]
 
 [dependencies]
-parity-wasm = "0.41"
-pwasm-utils = "0.12"
-wasmer-runtime = "0.17"
-wasmer-runtime-core = "0.17"
-assert_matches = "1.3.0"
-hex = "0.4"
-tempfile = "3.1.0"
+owasm = { path = "../owasm" }
 
 [profile.release]
 opt-level = 3
@@ -23,6 +17,6 @@ rpath = true
 lto = false
 debug-assertions = false
 codegen-units = 16
-panic = 'unwind'
+panic = "unwind"
 incremental = true
 overflow-checks = true
