+++ go-owasm/Makefile
@@ -5,16 +5,16 @@ USER_ID := $(shell id -u)
 USER_GROUP = $(shell id -g)
 
 docker-image-linux:
-	docker build . -t owasm/go-ext-builder:$(DOCKER_TAG)-linux -f ./Dockerfile.linux
+	DOCKER_BUILDKIT=1 docker build .. -t owasm/go-ext-builder:$(DOCKER_TAG)-linux -f Dockerfile.linux
 
 docker-image-osx:
-	docker build . -t owasm/go-ext-builder:$(DOCKER_TAG)-osx -f ./Dockerfile.osx
+	DOCKER_BUILDKIT=1 docker build .. -t owasm/go-ext-builder:$(DOCKER_TAG)-osx -f Dockerfile.osx
 
 docker-images: docker-image-linux docker-image-osx
 
 # and use them to compile release builds
 release:
 	rm -rf target/release
-	docker run --rm -u $(USER_ID):$(USER_GROUP) -v $(shell pwd):/code owasm/go-ext-builder:$(DOCKER_TAG)-osx
+	docker run --rm -u $(USER_ID):$(USER_GROUP) -v $(shell pwd)/..:/code owasm/go-ext-builder:$(DOCKER_TAG)-osx
 	rm -rf target/release
-	docker run --rm -u $(USER_ID):$(USER_GROUP) -v $(shell pwd):/code owasm/go-ext-builder:$(DOCKER_TAG)-linux
+	docker run --rm -u $(USER_ID):$(USER_GROUP) -v $(shell pwd)/..:/code owasm/go-ext-builder:$(DOCKER_TAG)-linux
