+++ go-owasm/cbindgen.toml
@@ -125,8 +125,8 @@ bitflags = false
 ############## Options for How Your Rust library Should Be Parsed ##############
 
 [parse]
-parse_deps = false
-# include = []
+parse_deps = true
+include = ["owasm"]
 exclude = []
 clean = false
 extra_bindings = []
