+++ owasm/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "owasm"
-version = "0.1.7"
+version = "0.1.8"
 authors = ["Band Protocol <connect@bandprotocol.com>"]
 edition = "2018"
 license = "Apache-2.0"
@@ -10,5 +10,12 @@ TODO
 """
 
 [dependencies]
+assert_matches = "1.3.0"
+hex = "0.4"
 json = "0.12.0"
 num = "*"
+parity-wasm = "0.41"
+pwasm-utils = "0.12"
+wasmer-runtime = "0.17"
+wasmer-runtime-core = "0.17"
+tempfile = "3.1.0"
