+++ owasm/src/lib.rs
@@ -29,5 +29,6 @@
 #[macro_use]
 mod macros;
 
+pub mod core;
 pub mod ext;
 pub mod oei;
